srtp msvc updates and cleanup
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@1087 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/srtp/crypto/cipher/aes_icm.c
@@ -208,7 +208,7 @@ aes_icm_set_octet(aes_icm_ctx_t *c,
 							  ((high32(octet_num) & 0x0f)<<(32-4)) |
 							   (low32(octet_num) >> 4));
 #else
-  int tail_num       = octet_num % 16;
+  int tail_num       = (int)(octet_num % 16);
   uint64_t block_num = octet_num / 16;
 #endif
   
@@ -329,7 +329,7 @@ aes_icm_encrypt_ismacryp(aes_icm_ctx_t *c,
               unsigned char *buf, unsigned int *enc_len, 
               int forIsmacryp) {
   unsigned int bytes_to_encr = *enc_len;
-  int i;
+  unsigned int i;
   uint32_t *b;
 
   /* check that there's enough segment left but not for ismacryp*/
