+++ libs/srtp/crypto/rng/prng.c
@@ -96,10 +96,10 @@ x917_prng_get_octet_string(uint8_t *dest, uint32_t len) {
   x917_prng.octet_count += len;
   
   /* find out the time */
-  t = time(NULL);
+  t = (uint32_t)time(NULL);
   
   /* loop until we have output enough data */
-  for (i=0; i < len/16; i++) {
+  for (i=0; (uint32_t)i < len/16; i++) {
     
     /* exor time into state */
     x917_prng.state.v32[0] ^= t;
